challenges/example/pyproject.toml
[tool.poetry]
name = "api-threats"
version = "0.1.0"
description = "API Threats website"
authors = ["Jose Haro Peralta"]
license = "MIT"
readme = "README.md"
#packages = [{include = "securing_apis_book"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
sqlalchemy = "^2.0.28"
psycopg = "^3.1.18"
requests = "^2.31.0"
python-jose = "^3.3.0"
uvicorn = "^0.28.0"
alembic = "^1.13.1"


[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
pre-commit = "^3.6.2"
schemathesis = "^3.31.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
